refactor(geometry): stand the point operators' device executors apart
`SigilGeometryMeshPopDevice` is now its own target, the way the mesh
renderer's device executor is, so a consumer that cooks a chain on the
host acquires no Diligent, no Vulkan loader and no Graphite. The three
entry points move to headers of their own under
`include/sigilgeometry/mesh/pop/device/`, and the conformance suites move
beside the sources they judge.

Beside it, what the report asked of the seam and its documents: the pop
device executor falls back to the host when the device refuses the
kernel, as the stamping and swept executors already did; the two
numberings of the builtin lanes become one; the staging texture a
readback crosses through is made once and grown; `deformFrame` becomes
internal and `seedCustomNames` goes, having no consumer; `shapers::wave`
answers both seams from the kit rather than growing `path::profile`;
and the comments and the README stop naming files that do not exist,
a test, and a consumer this library must not know about.

+++ apps/spell-circle-canvas/src/common/geometry/include/sigilgeometry/mesh/pop/Points.h
@@ -9,7 +9,7 @@
  *
  *  - instance()/quads(): stamp a Mesh (or a quad) onto every point —
  *    scale/tint/orientation read from lanes — producing ONE merged
- *    Mesh for render::drawMesh or world::World. "Instance planes
+ *    Mesh for render::drawMesh or for a 3D set. "Instance planes
  *    across points" is quads() + a normal lane (or leave normals off
  *    and let billboarding face the camera at draw time).
  *  - drawBillboards(): the UI-particle path — camera-facing sprites
@@ -109,9 +109,12 @@ void jitter(Cloud& cloud, float amplitude, uint32_t seed = 7);
 
 /** Smooth sin-field displacement (the organic drift) — the `pop::Noise`
  *  operator reached for without a chain, reading the same field through
- *  `pop::noiseField`. */
+ *  `pop::noiseField`. The seed is a FLOAT, as the operator's is: the
+ *  field reads it as one, and a modifier that could not reach a
+ *  fractional seed would be a narrower verb than the operator it is the
+ *  same verb as. */
 void displaceNoise(Cloud& cloud, float amplitude, float frequency,
-                   uint32_t seed = 7);
+                   float seed = 7);
 
 // ---------------------------------------------------------------------------
 // Consumers
